Add shell function wrapper to wm cli
As a consequence, the command is invoked with $wm instead of wm so we
can intercept the command with a shell function.

This was in preparation for adding the ability to toggle sending
graphical programs to the host machine, but it got a bit complex so it's
just disabled for now.